[GB] make scientificToDecimal a global helper
