+++ exchange/exchangeUtils.js
@@ -89,8 +89,36 @@ const isValidOrder = ({api, market, amount, price}) => {
   }
 }
 
+
+// https://gist.github.com/jiggzson/b5f489af9ad931e3d186
+const scientificToDecimal = num => {
+  if(/\d+\.?\d*e[\+\-]*\d+/i.test(num)) {
+    const zero = '0';
+    const parts = String(num).toLowerCase().split('e'); // split into coeff and exponent
+    const e = parts.pop(); // store the exponential part
+    const l = Math.abs(e); // get the number of zeros
+    const sign = e/l;
+    const coeff_array = parts[0].split('.');
+    if(sign === -1) {
+      num = zero + '.' + new Array(l).join(zero) + coeff_array.join('');
+    } else {
+      const dec = coeff_array[1];
+      if(dec) {
+        l = l - dec.length;
+      }
+      num = coeff_array.join('') + new Array(l+1).join(zero);
+    }
+  } else {
+    // make sure we always cast to string
+    num = num + '';
+  }
+
+  return num;
+}
+
 module.exports = {
   retry: retryInstance,
   bindAll,
-  isValidOrder
+  isValidOrder,
+  scientificToDecimal
 }
