+++ exchange/wrappers/binance.js
@@ -3,7 +3,9 @@ const _ = require('lodash');
 
 const Errors = require('../exchangeErrors');
 const marketData = require('./binance-markets.json');
-const retry = require('../exchangeUtils').retry;
+const exchangeUtils = require('../exchangeUtils');
+const retry = exchangeUtils.retry;
+const scientificToDecimal = exchangeUtils.scientificToDecimal;
 
 const Binance = require('binance');
 
@@ -264,36 +266,10 @@ Trader.prototype.round = function(amount, tickSize) {
   amount /= precision;
 
   // https://gist.github.com/jiggzson/b5f489af9ad931e3d186
-  amount = this.scientificToDecimal(amount);
+  amount = scientificToDecimal(amount);
 
   return amount;
 };
-
-// https://gist.github.com/jiggzson/b5f489af9ad931e3d186
-Trader.prototype.scientificToDecimal = function(num) {
-  if(/\d+\.?\d*e[\+\-]*\d+/i.test(num)) {
-    const zero = '0';
-    const parts = String(num).toLowerCase().split('e'); // split into coeff and exponent
-    const e = parts.pop(); // store the exponential part
-    const l = Math.abs(e); // get the number of zeros
-    const sign = e/l;
-    const coeff_array = parts[0].split('.');
-    if(sign === -1) {
-      num = zero + '.' + new Array(l).join(zero) + coeff_array.join('');
-    } else {
-      const dec = coeff_array[1];
-      if(dec) {
-        l = l - dec.length;
-      }
-      num = coeff_array.join('') + new Array(l+1).join(zero);
-    }
-  } else {
-    // make sure we always cast to string
-    num = num + '';
-  }
-
-  return num;
-}
 
 Trader.prototype.roundAmount = function(amount) {
   return this.round(amount, this.market.minimalOrder.amount);
@@ -405,7 +381,7 @@ Trader.prototype.getOrder = function(order, callback) {
   const reqData = {
     symbol: this.pair,
     // if this order was not part of the last 500 trades we won't find it..
-    limit: 500,
+    limit: 1000,
   };
 
   const handler = cb => this.binance.myTrades(reqData, this.handleResponse('getOrder', cb));
